+++ src/components/experience/BackButton/index.ts
@@ -1,0 +1,2 @@
+export * from './BackButton';
+export { default } from './BackButton'; // Default export
