+++ src/components/experience/DetailSection/DetailSection.scss
@@ -1,0 +1,22 @@
+@use "../../../App.scss"; // Adjusted path
+
+// Styles for DetailSection component
+.detail-section {
+    margin-bottom: 1.5rem;
+
+    h3 {
+        font-size: 1.25rem;
+        margin-bottom: 0.75rem;
+        color: white;
+    }
+
+    ul {
+        list-style-type: disc;
+        padding-left: 1.25rem;
+
+        li {
+            color: #d1d5db; // Light gray text
+            margin-bottom: 0.5rem;
+        }
+    }
+}
