+++ src/components/experience/DetailSection/index.ts
@@ -1,0 +1,2 @@
+export * from './DetailSection';
+export { default } from './DetailSection'; // Default export
