+++ src/components/experience/ExpandedCardOverlay/index.ts
@@ -1,0 +1,2 @@
+export * from './ExpandedCardOverlay';
+export { default } from './ExpandedCardOverlay'; // Default export
