+++ src/components/experience/ExperienceList/index.ts
@@ -1,0 +1,2 @@
+export * from './ExperienceList';
+export { default } from './ExperienceList'; // Default export
