+++ src/components/experience/JobCard/JobCard.tsx
@@ -1,9 +1,10 @@
 import React, { useRef } from "react";
-import TimeFrame from "./TimeFrame";
-import JobTitle from "./JobTitle";
-import JobDescription from "./JobDescription";
-import SkillTags from "./SkillTags";
-import { JobCardProps } from "../../types/experience";
+import TimeFrame from "../TimeFrame"; // Adjusted path
+import JobTitle from "../JobTitle"; // Adjusted path
+import JobDescription from "../JobDescription"; // Adjusted path
+import SkillTags from "../SkillTags"; // Adjusted path
+import { JobCardProps } from "../../../types/experience"; // Adjusted path
+import './JobCard.scss'; // New SCSS import
 
 const JobCard: React.FC<JobCardProps> = ({
     timeRange,
