+++ src/components/experience/JobCard/index.ts
@@ -1,0 +1,2 @@
+export * from './JobCard';
+export { default } from './JobCard'; // Default export
