+++ src/components/experience/JobDescription/JobDescription.scss
@@ -1,0 +1,8 @@
+@use "../../../App.scss"; // Adjusted path
+
+// Styles for JobDescription component
+.job-description {
+    font-size: 1.125rem;
+    margin-bottom: 1.5rem;
+    color: #d1d5db; // Light gray text
+}
