+++ src/components/experience/JobDescription/JobDescription.tsx
@@ -1,5 +1,6 @@
 import React from "react";
-import { JobDescriptionProps } from "../../types/experience";
+import { JobDescriptionProps } from "../../../types/experience"; // Adjusted path
+import './JobDescription.scss'; // New SCSS import
 
 const JobDescription: React.FC<JobDescriptionProps> = ({ description }) => (
     <p className="job-description">
