+++ src/components/experience/JobDescription/index.ts
@@ -1,0 +1,2 @@
+export * from './JobDescription';
+export { default } from './JobDescription'; // Default export
