+++ src/components/experience/JobTitle/index.ts
@@ -1,0 +1,2 @@
+export * from './JobTitle';
+export { default } from './JobTitle'; // Default export
