+++ src/components/experience/SkillTag/SkillTag.tsx
@@ -1,5 +1,6 @@
 import React from "react";
-import { SkillTagProps } from "../../types/experience";
+import { SkillTagProps } from "../../../types/experience"; // Adjusted path
+import './SkillTag.scss'; // New SCSS import
 
 const SkillTag: React.FC<SkillTagProps> = ({ skill }) => (
     <span className="skill-tag">
