+++ src/components/experience/SkillTag/index.ts
@@ -1,0 +1,2 @@
+export * from './SkillTag';
+export { default } from './SkillTag'; // Default export
