+++ src/components/experience/SkillTags/SkillTags.scss
@@ -1,0 +1,8 @@
+@use "../../../App.scss"; // Adjusted path
+
+// Styles for SkillTags component
+.skill-tags {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 0.5rem;
+}
