+++ src/components/experience/SkillTags/SkillTags.tsx
@@ -1,6 +1,7 @@
 import React from "react";
-import SkillTag from "./SkillTag";
-import { SkillTagsProps } from "../../types/experience";
+import SkillTag from "../SkillTag"; // Adjusted path
+import { SkillTagsProps } from "../../../types/experience"; // Adjusted path
+import './SkillTags.scss'; // New SCSS import
 
 const SkillTags: React.FC<SkillTagsProps> = ({ skills }) => (
     <div className="skill-tags">
