+++ src/components/experience/SkillTags/index.ts
@@ -1,0 +1,2 @@
+export * from './SkillTags';
+export { default } from './SkillTags'; // Default export
