+++ src/components/experience/TimeFrame/TimeFrame.scss
@@ -1,0 +1,16 @@
+@use "../../../App.scss"; // Adjusted path
+
+// Styles for TimeFrame component
+.time-frame {
+    margin-bottom: 1rem;
+
+    @media (min-width: App.$mobile-view-transition-threshold) {
+        width: 20%;
+        margin-bottom: 0;
+    }
+
+    p {
+        font-size: 1.125rem;
+        color: #9ca3af; // Gray text
+    }
+}
