+++ src/components/experience/TimeFrame/TimeFrame.tsx
@@ -1,5 +1,6 @@
 import React from "react";
-import { TimeFrameProps } from "../../types/experience";
+import { TimeFrameProps } from "../../../types/experience"; // Adjusted path
+import './TimeFrame.scss'; // New SCSS import
 
 const TimeFrame: React.FC<TimeFrameProps> = ({ timeRange }) => (
     <div className="time-frame">
