+++ src/components/experience/TimeFrame/index.ts
@@ -1,0 +1,2 @@
+export * from './TimeFrame';
+export { default } from './TimeFrame'; // Default export
