+++ src/components/experience/index.ts
@@ -1,14 +1,23 @@
-import ExperienceList from "./ExperienceList";
-
+// Default export for the entire experience module
+import ExperienceList from './ExperienceList'; // Imports from ./ExperienceList/index.ts
 export default ExperienceList;
 
 // Export individual components for reuse elsewhere if needed
-export { default as TimeFrame } from "./TimeFrame";
-export { default as JobTitle } from "./JobTitle";
-export { default as JobDescription } from "./JobDescription";
-export { default as SkillTag } from "./SkillTag";
-export { default as SkillTags } from "./SkillTags";
-export { default as DetailSection } from "./DetailSection";
-export { default as BackButton } from "./BackButton";
-export { default as ExpandedCardOverlay } from "./ExpandedCardOverlay";
-export { default as JobCard } from "./JobCard";
+// These will also import from the respective component's sub-directory index.ts
+export { default as BackButton } from './BackButton';
+export { default as DetailSection } from './DetailSection';
+export { default as ExpandedCardOverlay } from './ExpandedCardOverlay';
+// ExperienceList is already default, but can be named exported too if desired
+export { default as ExperienceList } from './ExperienceList';
+export { default as JobCard } from './JobCard';
+export { default as JobDescription } from './JobDescription';
+export { default as JobTitle } from './JobTitle';
+export { default as SkillTag } from './SkillTag';
+export { default as SkillTags } from './SkillTags';
+export { default as TimeFrame } from './TimeFrame';
+
+// Optionally, if you want to allow importing all specific types/interfaces from components:
+// export * from './BackButton';
+// export * from './DetailSection';
+// ...etc.
+// For now, just exporting the default components as named exports.
